Put WebJob Sample in a .NetCore Project
