+++ GeekLearning.Events.sln
@@ -25,7 +25,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GeekLearning.Events.Sample"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GeekLearning.Events.SampleEventModels", "samples\GeekLearning.Events.SampleEventModels\GeekLearning.Events.SampleEventModels.csproj", "{B7D86F8D-7E35-464F-BE2A-97903FD38719}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GeekLearning.Events.WebJobSample", "samples\GeekLearning.Events.WebJobSample\GeekLearning.Events.WebJobSample.csproj", "{C2CF4813-4DFC-41B0-A75E-3E52359B1028}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GeekLearning.Events.SampleWebJob", "samples\GeekLearning.Events.SampleWebJob\GeekLearning.Events.SampleWebJob.csproj", "{C1454913-6CF1-4BA1-A2DF-E9E78B4FA507}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,10 +53,10 @@ Global
 		{B7D86F8D-7E35-464F-BE2A-97903FD38719}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B7D86F8D-7E35-464F-BE2A-97903FD38719}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B7D86F8D-7E35-464F-BE2A-97903FD38719}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C2CF4813-4DFC-41B0-A75E-3E52359B1028}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C2CF4813-4DFC-41B0-A75E-3E52359B1028}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C2CF4813-4DFC-41B0-A75E-3E52359B1028}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C2CF4813-4DFC-41B0-A75E-3E52359B1028}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C1454913-6CF1-4BA1-A2DF-E9E78B4FA507}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1454913-6CF1-4BA1-A2DF-E9E78B4FA507}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1454913-6CF1-4BA1-A2DF-E9E78B4FA507}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1454913-6CF1-4BA1-A2DF-E9E78B4FA507}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -67,7 +67,7 @@ Global
 		{40A0CD11-E804-45DB-B7B5-5D58D37D28A6} = {1A754E40-F164-4442-B51E-160CF2A1F34E}
 		{D5DAA9A0-CD8D-473D-8FDF-0350A7EEEE78} = {31AED30E-FF14-4D84-AB6E-61DD8889E50C}
 		{B7D86F8D-7E35-464F-BE2A-97903FD38719} = {31AED30E-FF14-4D84-AB6E-61DD8889E50C}
-		{C2CF4813-4DFC-41B0-A75E-3E52359B1028} = {31AED30E-FF14-4D84-AB6E-61DD8889E50C}
+		{C1454913-6CF1-4BA1-A2DF-E9E78B4FA507} = {31AED30E-FF14-4D84-AB6E-61DD8889E50C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E09B2C30-A610-4720-BFA8-DB61F944C041}
