+++ samples/GeekLearning.Events.Sample/GeekLearning.Events.Sample.csproj
@@ -20,7 +20,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\GeekLearning.Events.SampleEventModels\GeekLearning.Events.SampleEventModels.csproj" />
     <ProjectReference Include="..\..\src\GeekLearning.Events.InMemory\GeekLearning.Events.InMemory.csproj" />
     <ProjectReference Include="..\..\src\GeekLearning.Events.AzureStorage\GeekLearning.Events.AzureStorage.csproj" />
     <ProjectReference Include="..\GeekLearning.Events.SampleEventModels\GeekLearning.Events.SampleEventModels.csproj" />
