+++ samples/GeekLearning.Events.SampleWebJob/GeekLearning.Events.SampleWebJob.csproj
@@ -1,0 +1,24 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net462</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.PlatformAbstractions" Version="1.1.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions" Version="2.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\GeekLearning.Events\GeekLearning.Events.csproj" />
+    <ProjectReference Include="..\GeekLearning.Events.SampleEventModels\GeekLearning.Events.SampleEventModels.csproj" />
+  </ItemGroup>
+
+
+</Project>
