--- samples/GeekLearning.Events.WebJobSample/Program.cs
@@ -1,42 +1,0 @@
-﻿namespace GeekLearning.Events.WebJobSample
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.WebJobs;
-    using Microsoft.Extensions.DependencyInjection;
-
-    // To learn more about Microsoft Azure WebJobs SDK, please see https://go.microsoft.com/fwlink/?LinkID=320976
-    class Program
-    {
-        // Please set the following connection strings in app.config for this WebJob to run:
-        // AzureWebJobsDashboard and AzureWebJobsStorage
-        static void Main()
-        {
-            var services = new ServiceCollection();
-            ConfigureServices(services);
-            var serviceProvider = services.BuildServiceProvider();
-
-            var config = new JobHostConfiguration
-            {
-                JobActivator = new ServiceProviderJobActivator(serviceProvider)
-            };
-
-            if (config.IsDevelopment)
-            {
-                config.UseDevelopmentSettings();
-            }
-
-            var host = new JobHost(config);
-            // The following code ensures that the WebJob will be running continuously
-            host.RunAndBlock();
-        }
-
-        private static void ConfigureServices(ServiceCollection services)
-        {
-            services.AddEventReceiver();
-        }
-    }
-}
