--- samples/GeekLearning.Events.WebJobSample/packages.config
@@ -1,61 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Azure.KeyVault.Core" version="2.0.4" targetFramework="net462" />
-  <package id="Microsoft.Azure.WebJobs" version="2.0.0" targetFramework="net462" />
-  <package id="Microsoft.Azure.WebJobs.Core" version="2.0.0" targetFramework="net462" />
-  <package id="Microsoft.Data.Edm" version="5.8.3" targetFramework="net462" />
-  <package id="Microsoft.Data.OData" version="5.8.3" targetFramework="net462" />
-  <package id="Microsoft.Data.Services.Client" version="5.8.3" targetFramework="net462" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.0.0" targetFramework="net462" />
-  <package id="Microsoft.NETCore.Platforms" version="2.0.0" targetFramework="net462" />
-  <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.3" targetFramework="net462" />
-  <package id="NETStandard.Library" version="2.0.1" targetFramework="net462" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net462" />
-  <package id="System.AppContext" version="4.3.0" targetFramework="net461" />
-  <package id="System.Collections" version="4.3.0" targetFramework="net461" />
-  <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net461" />
-  <package id="System.ComponentModel" version="4.3.0" targetFramework="net461" />
-  <package id="System.Console" version="4.3.0" targetFramework="net461" />
-  <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="net461" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.4.1" targetFramework="net462" />
-  <package id="System.Diagnostics.Tools" version="4.3.0" targetFramework="net461" />
-  <package id="System.Diagnostics.Tracing" version="4.3.0" targetFramework="net461" requireReinstallation="true" />
-  <package id="System.Globalization" version="4.3.0" targetFramework="net461" />
-  <package id="System.Globalization.Calendars" version="4.3.0" targetFramework="net461" />
-  <package id="System.IO" version="4.3.0" targetFramework="net461" requireReinstallation="true" />
-  <package id="System.IO.Compression" version="4.3.0" targetFramework="net461" />
-  <package id="System.IO.Compression.ZipFile" version="4.3.0" targetFramework="net461" />
-  <package id="System.IO.FileSystem" version="4.3.0" targetFramework="net461" />
-  <package id="System.IO.FileSystem.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="System.Linq" version="4.3.0" targetFramework="net461" />
-  <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net461" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net462" />
-  <package id="System.Net.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="System.Net.Sockets" version="4.3.0" targetFramework="net461" />
-  <package id="System.ObjectModel" version="4.3.0" targetFramework="net461" />
-  <package id="System.Reflection" version="4.3.0" targetFramework="net461" requireReinstallation="true" />
-  <package id="System.Reflection.Extensions" version="4.3.0" targetFramework="net461" />
-  <package id="System.Reflection.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="net461" />
-  <package id="System.Runtime" version="4.3.0" targetFramework="net461" requireReinstallation="true" />
-  <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net461" requireReinstallation="true" />
-  <package id="System.Runtime.Handles" version="4.3.0" targetFramework="net461" />
-  <package id="System.Runtime.InteropServices" version="4.3.0" targetFramework="net461" requireReinstallation="true" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net461" />
-  <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.3.1" targetFramework="net462" />
-  <package id="System.Spatial" version="5.8.3" targetFramework="net462" />
-  <package id="System.Text.Encoding" version="4.3.0" targetFramework="net461" />
-  <package id="System.Text.Encoding.Extensions" version="4.3.0" targetFramework="net461" />
-  <package id="System.Text.RegularExpressions" version="4.3.0" targetFramework="net461" />
-  <package id="System.Threading" version="4.3.0" targetFramework="net461" />
-  <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net461" />
-  <package id="System.Threading.Timer" version="4.3.0" targetFramework="net461" />
-  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="net461" />
-  <package id="System.Xml.XDocument" version="4.3.0" targetFramework="net461" />
-  <package id="WindowsAzure.Storage" version="8.5.0" targetFramework="net462" />
-</packages>
